requirements.txt
chardet==5.2.0
cssselect2==0.7.0
lxml==5.2.2
mysql-connector==2.2.9
numpy==1.24.4
opencv-python==4.9.0.80
pdf2image==1.17.0
pillow==10.3.0
reportlab==4.2.0
svglib==1.5.1
tinycss2==1.3.0
webencodings==0.5.1
flask_cors==3.0.10
flask-mail==0.9.1

# sudo apt-get install mysql-server
# sudo apt install python3-tk -y
